Add program to crack passwords using rainbow table.
The password + salt has to be entered manually because the other
groupmates failed to produce a working password hasher.

The hash chain length and the filename of the rainbow table are
hard-coded into the program, so those can't be changed from
chain-length="16" and filename="rainbow-table-16.txt".

For creating the hash, the password is limited to 10 characters and the
salt is limited to 4095 max.